requirements.txt
torch
torchvision
pillow
numpy
matplotlib
scikit-learn
tqdm
